# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin-installment": {
            "get": {
                "description": "GetAdminInstallment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "GetAdminInstallment",
                "parameters": [
                    {
                        "description": "installment",
                        "name": "model.Installment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Installment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Installment"
                        }
                    }
                }
            }
        },
        "/all-installment": {
            "get": {
                "description": "GetAllInstallment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "installment"
                ],
                "summary": "GetAllInstallment",
                "parameters": [
                    {
                        "description": "installment",
                        "name": "model.Installment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Installment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Installment"
                        }
                    }
                }
            }
        },
        "/all-loan": {
            "get": {
                "description": "GetAllLoan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan"
                ],
                "summary": "GetAllLoan",
                "parameters": [
                    {
                        "description": "Loan",
                        "name": "model.Loan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Loan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Loan"
                        }
                    }
                }
            }
        },
        "/create-loan": {
            "post": {
                "description": "Create Loan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan"
                ],
                "summary": "Create Loan",
                "parameters": [
                    {
                        "description": "Loan",
                        "name": "model.Loan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Loan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Loan"
                        }
                    }
                }
            }
        },
        "/detail-loan/{userId}": {
            "get": {
                "description": "GetDetailLoan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan"
                ],
                "summary": "GetDetailLoan",
                "parameters": [
                    {
                        "description": "Loan",
                        "name": "model.Loan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Loan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Loan"
                        }
                    }
                }
            }
        },
        "/history-instalment/{userId}": {
            "get": {
                "description": "GetHistInstallment",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "installment"
                ],
                "summary": "GetHistInstallment",
                "parameters": [
                    {
                        "description": "installment",
                        "name": "model.Installment",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Installment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Installment"
                        }
                    }
                }
            }
        },
        "/update/{Id}": {
            "put": {
                "description": "Update Loan",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "loan"
                ],
                "summary": "Update Loan",
                "parameters": [
                    {
                        "description": "Loan",
                        "name": "model.Loan",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Loan"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Loan"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Installment": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "loan_id": {
                    "type": "integer"
                },
                "nominal": {
                    "type": "integer"
                },
                "payment_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.Loan": {
            "type": "object",
            "properties": {
                "cost_installment": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "item": {
                    "type": "string"
                },
                "nominal": {
                    "type": "integer"
                },
                "paid": {
                    "type": "integer"
                },
                "period_installment": {
                    "type": "integer"
                },
                "remain_installment": {
                    "type": "integer"
                },
                "status_loan": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    }
}